Created module structure for GridParser
