+++ src/lib/GridParser/ColumnDefs/ErrorGrid/index.js
@@ -1,0 +1,8 @@
+/*
+Copies all fields and header names from processed form and 
+return error row?
+
+
+Stringify values with error message.
+
+*/
