+++ src/lib/GridParser/ColumnDefs/HideColumns/index.js
@@ -1,0 +1,11 @@
+/*
+var columnDefs = [
+    {
+       headerName: "Stone_ID",
+       field: "Stone_ID",
+       width: 100,
+       hide: true,
+       suppressToolPanel: true
+    }
+]
+*/
