+++ src/lib/GridParser/ColumnDefs/IsValid/index.js
@@ -1,0 +1,6 @@
+//validator
+//is_valid
+//and is_error
+
+function IsValid(){}
+function IsError(){}
