+++ src/lib/GridParser/ColumnDefs/README
@@ -1,0 +1,1 @@
+Creates configurations for aggrid.
