+++ src/lib/GridParser/InitialParams/README
