+++ src/lib/GridParser/InitialParams/index.js
@@ -1,0 +1,2 @@
+//how to handle rows
+//i.e. new row, update, delete, etc
