--- src/lib/GridParser/ValueTypes/TypeTransform.js
@@ -1,11 +1,0 @@
-//set everything to string unless null?
-
-class ParseGrid {
-    constructor (json_object) {
-        this.jx = json_object
-    }
-    AssembleColumnDefs() {
-        
-    }
-
-}
