--- src/lib/InitialParams/MetaField/index.js
@@ -1,3 +1,0 @@
-//meta field
-//meta field name
-//meta fiel backup
